dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "MQTT backend + dashboard for Whacamole game"
readme = "README.md"
license-file = "../LICENSE"
authors = [{ name = "darragh0" }, { name = "jjola00" }, { name = "tomjoyce1" }]
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.124.0",
  "paho-mqtt>=2.1.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.2.1",
  "python-multipart>=0.0.20",
  "uvicorn>=0.38.0",
]

[project.scripts]
dashboard = "dashboard.__main__:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.pyright]
pythonVersion = "3.12"
